Compile fix for Unicode build on win64
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@35913 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/variant.cpp
@@ -706,7 +706,7 @@ bool wxVariantDataChar::Read(wxInputStream& str)
 
 bool wxVariantDataChar::Read(wxString& str)
 {
-    m_value = str.ToAscii()[0u];
+    m_value = str.ToAscii()[size_t(0)];
     return true;
 }
